tambah opsi cooldown dan ambang confidence untuk adapter SDWS
